config, eslint, file upload, api changes, packages

+++ index.js
@@ -1,47 +1,66 @@
+require("dotenv").config();
+const { PORT, DOMAIN, TOKEN, CHARS, IDLENGTH, DELETELENGTH } = process.env;
 const express = require("express");
 const multer = require("multer");
 const helmet = require("helmet");
 const Enmap = require("enmap");
-const db = new Enmap("deletes");
-const upload = multer({ dest: "images/" });
+const db = new Enmap("files");
+const upload = multer({ dest: "tmp/" });
 const { customAlphabet } = require("nanoid");
-const chars = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz";
-const genId = customAlphabet(chars, 8);
-const genDeleteId = customAlphabet(chars, 32);
+const genId = customAlphabet(CHARS, parseInt(IDLENGTH));
+const genDeleteToken = customAlphabet(CHARS, parseInt(DELETELENGTH));
 const fs = require("fs");
 const app = new express();
-require("dotenv").config();
 
+if(!fs.existsSync("files")) fs.mkdirSync("files");
 if(!fs.existsSync("images")) fs.mkdirSync("images");
 
-app.listen(process.env.PORT, () => {
-    console.log("Webserver online!");
+app.listen(PORT, () => {
+    console.log(`Webserver running on port ${PORT}, ` +
+                `accesible at https://${DOMAIN}/`);
 });
 
 app.use(helmet());
 
 app.get("/", (_req, res) => {
     res.status(200).sendFile(__dirname + "/index.html");
-})
+});
 
-app.use(express.static("images"));
+app.use("/i/", express.static("images"));
+
+app.get("/f/:id", (req, res) => {
+    let id = req.params.id.split(".")[0];
+    if(!db.has(id)) return res.sendStatus(404);
+    let { file, name } = db.get(id);
+    res.download(file, name);
+});
 
 const authenticate = (req, res, next) => {
-    if(req.query.token != process.env.TOKEN) return res.sendStatus(401);
+    if(req.query.token != TOKEN) return res.sendStatus(401);
     else next();
-}
+};
 
-app.post("/upload", authenticate, upload.single("image"), (req, res) => {
+app.post("/upload", authenticate, upload.single("file"), (req, res) => {
     try {
-        let name = genId() + "." + req.file.originalname.split(".")[1];
-        let deleteId = genDeleteId();
-        fs.rename(req.file.path, `images/${name}`, () => {
-            db.set(deleteId, name);
+        if(!(["file", "image"].includes(req.query.type)))
+            return res.sendStatus(400);
+        let isImage = req.query.type == "image";
+        let ext = "." + req.file.originalname.split(".").reverse()[0];
+        let id = genId();
+        let deleteToken = genDeleteToken();
+        let file = id + (ext != "." && isImage ? ext : "");
+        let path = `${isImage ? "images" : "files"}/${file}`;
+        fs.rename(req.file.path, path, () => {
+            db.set(id, {
+                file: path,
+                name: req.file.originalname,
+                token: deleteToken
+            });
             res.status(200).send({
                 status: 200,
-                url: `https://i.antti.codes/${name}`,
-                delete: `https://i.antti.codes/delete/${deleteId}`
-            })
+                url: `https://${DOMAIN}/${isImage ? "i" : "f"}/${file}`,
+                delete: `https://${DOMAIN}/d/${id}/${deleteToken}`
+            });
         });
     } catch(e) {
         console.error(e);
@@ -50,17 +69,17 @@ app.post("/upload", authenticate, upload.single("image"), (req, res) => {
             message: "Internal server error occured trying to upload file"
         });
     }
-})
+});
 
-app.get("/delete/:id", (req, res) => {
+app.get("/d/:id/:token", (req, res) => {
     try {
-        fs.unlink(`./images/${db.get(req.params.id)}`, () => {
-            res.status(200).send({
-                status: 200,
-                message: "File delete succesfully."
-            });
+        if(!db.has(req.params.id)) return res.sendStatus(404);
+        let { token, file } = db.get(req.params.id);
+        if(token != req.params.token) return res.sendStatus(401);
+        fs.unlink("./" + file, () => {
+            res.sendStatus(200);
+            db.delete(req.params.id);
         });
-        db.delete(req.params.id);
     } catch (e) {
         console.error(e);
         res.status(500).send({
@@ -68,4 +87,4 @@ app.get("/delete/:id", (req, res) => {
             message: "Internal server error occured trying to delete the file"
         });
     }
-})
+});
